load-reviews/requirements.txt
google-api-python-client
grpcio
firebase-admin
python-telegram-bot
